+++ apps/nowait-user/src/App.tsx
@@ -1,7 +1,7 @@
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import Router from "./routes/Router";
 import { BrowserRouter } from "react-router-dom";
-import GlobalToast from "./components/common/Toast";
+import GlobalToast from "./components/common/toast/Toast";
 import 'react-toastify/dist/ReactToastify.css';
 
 function App() {
